Fix type mismatches

+++ soem/ethercatprint.h
@@ -52,7 +52,7 @@ extern "C"
 {
 #endif
 
-char* ec_sdoerror2string( uint16 sdoerrorcode);
+char* ec_sdoerror2string( uint32 sdoerrorcode);
 char* ec_ALstatuscode2string( uint16 ALstatuscode);
 char* ec_soeerror2string( uint16 errorcode);
 char* ecx_elist2string(ecx_contextt *context);
